Rover: move update_sensor_flags to GCS_Rover.cpp
